Code refactoring in the visualization tool
