+++ src/gui/gui_colors.h
@@ -1,5 +1,5 @@
 #ifndef __GUI_COLORS_H
-#define __GUI_COLORS_H 
+#define __GUI_COLORS_H
 
 #include "../3dparty/raylib/src/raylib.h"
 #define NUM_SCALES 7
